Remove visual studio files from the repository.
I wrote a C# utility called VsProjectGenerator which will generate the VS solution and projects from a simple config file. The VsProjectGenerator should now be the only visual studio project that ever exists in the repository, and is isolated from the rest of the source code.
Also wrote a basic MIDI interface using RtMidi which prints incoming MIDI data to the console.