Add the ability to create constant nodes from within a native module call.
Native module libraries now have hooks to run custom code when the compiler is initialized and deinitialized. Task function libraries have hooks for when the engine is initialized and deinitialized as well as hooks to run a function directly before and after task functions are initialized. This allows libraries to be fully self-contained (e.g. c_sample_library can be moved into the sampler library).