Added automatic latency compensation. Native modules can report latency via the get_latency callback. Latency can be queried and manually added in scripts using core.get_latency() and core.add_latency(), respectively. In addition, native modules have an optional validate_arguments callback which always runs at compile-time and can report errors which previously would have only been detected at runtime (e.g. a negative delay value).
The delay functions have been rewritten from scratch in a more robust way that allows for constant detection. Delay is implemented for both reals and bools using either samples or seconds as the delay value input and an optional initial value can be specified as well. Note: delay has not been tested, bugs will likely be fixed in a follow-up change.